backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
